updates for L24.11
TILER allocations that used to be proxied via syslink_daemon now are
handled by the linux side client by way of a memsrv RcmServer.  This causes
any allocations done by codec to be recovered when the libdce process
exits.

It would be nicer if syslink provided a memsrv via a shared library so we
didn't have to copy all that code into this tree.. but until then, memsrv.c
is a hacked up variant of one of the syslink sample apps.