remove test
Moving test to omapdrmtest tree, to be able to re-use display and other
utility code for displaying video directly with drm plane (overlay) or
with dri2 video.

The v4l2/camera util code and fill pattern generation should also make
it easier to add an encoder test in the future.