ducati: dce: Remove resource table trace print from image
In order to have the linker include the resource table in the
baseimage, the resource variable needs to be used in some form.
Otherwise the optimization in the linker detects that this is a
non-referenced variable and removes it from the baseimage. To
alleviate this, a dummy trace was used to make sure the resource
table section is present in the firmware binary.

The new IpcMemory module inherently adds this trace, and so the
trace in DCE image is no longer required and is removed.

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>