ducati: split ti.dce into a lib and a separate baseimage
The 'package' ti.dce both had the DCE code, as well as the
baseimage code (BIOS cfg, main(), ...). With this change we split the
DCE code from the baseimage. there is now:
 - ti.dce
 - ti.dce.baseimage

that will allow a simpler reuse of ti.dce into other base images

there is still some clean up that can be done in the base image to move
the thread creation, and some config in ti.dce, and this will be done
in further commits. This commit only changes the file 'location' without
changing any implementation detail.

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>